+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add `String#isUpperCase()` and `String#isLowerCase()`
 * Add `ignore_arguments` option to the `memoize()` decorator
 * Add `Object#zip(keys, values)` and `Object#unzip(obj)`
+* Add `Blast.checksumSymbol` you can use to define a method that replaces the parameter of a `Object.checksum(obj)` call
 
 ## 0.6.4 (2019-01-12)