+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Make the internal `queueImmediate`, `queueTick` methods asynchronous even after Blast had loaded
 * `Blast.doLoaded()` will now also work after Protoblast has already loaded once
 * Allow passing a regex into `RegExp.interpret()`
+* Add a `main_class` getter to namespaces
 
 ## 0.6.4 (2019-01-12)